Move all distance-related code to distances.py
