+++ co2calculator/distances.py
@@ -4,24 +4,70 @@
 """Functions for obtaining the distance between given addresses."""
 
 
-from typing import Tuple
-from ._types import Kilometer
+from typing import Tuple, Union, Optional
+import enum
+from pathlib import Path
+from dotenv import load_dotenv
+import os
+import warnings
+
+
 import numpy as np
+import pandas as pd
 import openrouteservice
 from openrouteservice.directions import directions
 from openrouteservice.geocode import pelias_search, pelias_structured
-import os
-from pathlib import Path
-from dotenv import load_dotenv
-import pandas as pd
 from thefuzz import fuzz
 from thefuzz import process
-import warnings
+from pydantic import BaseModel
+
+
+from ._types import Kilometer
+
 
 load_dotenv()  # take environment variables from .env.
 
+# Load environment vars (TODO: Use pydantic.BaseSettings)
 ORS_API_KEY = os.environ.get("ORS_API_KEY")
+
+# Set (module) global vars (TODO: Don't do it - make it a class and move it to attributes!)
 script_path = str(Path(__file__).parent)
+detour_df = pd.read_csv(f"{script_path}/../data/detour.csv")
+
+# Module's models
+@enum.unique
+class TransportationMode(enum.Enum):
+    CAR = enum.auto()
+    MOTORBIKE = enum.auto()
+    BUS = enum.auto()
+    TRAIN = enum.auto()
+    PLANE = enum.auto()
+    FERRY = enum.auto()
+
+
+class StructuredLocation(BaseModel):
+    address: Optional[str]
+    locality: str
+    country: str
+
+
+class TrainStation(BaseModel):
+    station_name: str
+    country: str  # NOTE: Could be improved with validator to country codes
+
+
+class DistanceRequest(BaseModel):
+    transportation_mode: TransportationMode
+    start: Union[StructuredLocation, TrainStation]
+    destination: Union[StructuredLocation, TrainStation]
+
+
+IataCode = str  # NOTE: Could be improved with validation of IATA codes
+
+
+# Module's exceptions
+class InvalidSpatialInput(Exception):
+    """Raised when consumer inputs invalid spatial information"""
 
 
 def haversine(
@@ -308,3 +354,146 @@ def get_route(coords, profile: str = None) -> Kilometer:
     )  # divide my 1000, as we're working with distances in km
 
     return dist
+
+
+def _apply_detour(distance: Kilometer, transportation_mode: str) -> Kilometer:
+    """
+    Function to apply specific detour parameters to a distance as the crow flies
+    :param distance: Distance as the crow flies between location of departure and destination of a trip
+    :param transportation_mode: Mode of transport used in the trip
+    :type distance: float
+    :type transportation_mode: str
+    :return: Distance accounted for detour
+    :rtype: float
+    """
+    try:
+        detour_coefficient = detour_df[
+            detour_df["transportation_mode"] == transportation_mode
+        ]["coefficient"].values[0]
+        detour_constant = detour_df[
+            detour_df["transportation_mode"] == transportation_mode
+        ]["constant [km]"].values[0]
+    except KeyError:
+        detour_coefficient = 1.0
+        detour_constant = 0.0
+        warnings.warn(
+            f"""
+        No detour coefficient or constant available for this transportation mode.
+        Detour parameters are available for the following transportation modes:
+        Using detour_coefficient = {detour_coefficient} and detour_constant = {detour_constant}.
+        """
+        )
+    distance_with_detour = detour_coefficient * distance + detour_constant
+
+    return distance_with_detour
+
+
+def get_distance(start, destination, transportation_mode):
+    """Get the distance between start and destination
+
+    Raises:
+    - InvalidSpatialInput if start and stop are malformed or None
+    """
+
+    # TODO: Refactor to meet DRY
+
+    if None in [start, destination]:
+        raise InvalidSpatialInput("neither distance or start/destination  provided")
+
+    if transportation_mode == "car":
+        # Stops are formatted like:
+        # [
+        #   {
+        #     "address": "Im Neuenheimer Feld 348",
+        #     "locality": "Heidelberg",
+        #     "country": "Germany"
+        #   },
+        #   {
+        #     "country": "Germany",
+        #     "locality": "Berlin",
+        #     "address": "Alexanderplatz 1"
+        #   }
+        # ]
+        # TODO: Validate with BaseModel
+
+        coords = []
+        for loc in [start, destination]:
+            _, _, loc_coords, _ = geocoding_structured(loc)
+            coords.append(loc_coords)
+        return get_route(coords, "driving-car")
+
+    if transportation_mode == "motorbike":
+        # Same model as car!
+        # [
+        #     {"address": "Im Neuenheimer Feld 348", "locality": "Heidelberg", "country": "Germany"},
+        #     {"country": "Germany", "locality": "Berlin", "address": "Alexanderplatz 1"}
+        # ]
+        coords = []
+        for loc in [start, destination]:
+            _, _, loc_coords, _ = geocoding_structured(loc)
+            coords.append(loc_coords)
+        distance = get_route(coords, "driving-car")
+
+    if transportation_mode == "bus":
+        # Same as car (StructuredLocation)
+        # TODO: Validate with BaseModel
+        # TODO: Question: Why are we not calculating the bus trip like `driving-car` routes?
+
+        distance = 0
+        coords = []
+        for loc in [start, destination]:
+            _, _, loc_coords, _ = geocoding_structured(loc)
+            coords.append(loc_coords)
+        for i in range(0, len(coords) - 1):
+            distance += haversine(
+                coords[i][1], coords[i][0], coords[i + 1][1], coords[i + 1][0]
+            )
+        return _apply_detour(distance, transportation_mode)
+
+    if transportation_mode == "train":
+        # Stops for train distance calculations are formed like this:
+        # [
+        #     {"station_name": "Heidelberg Hbf", "country": "DE"},
+        #     {"station_name": "Berlin Hauptbahnhof", "country": "DE"}
+        # ]
+        # TODO: Validate with BaseModel
+
+        distance = 0
+        coords = []
+
+        for loc in [start, destination]:
+            try:
+                _, _, loc_coords = geocoding_train_stations(loc)
+            except RuntimeWarning:
+                _, _, loc_coords, _ = geocoding_structured(loc)
+            except ValueError:
+                _, _, loc_coords, _ = geocoding_structured(loc)
+            coords.append(loc_coords)
+
+        for i in range(len(coords) - 1):
+            # compute great circle distance between locations
+            distance += haversine(
+                coords[i][1], coords[i][0], coords[i + 1][1], coords[i + 1][0]
+            )
+        return _apply_detour(distance, transportation_mode)
+
+    if transportation_mode == "plane":
+        # Stops are IATA code of airports
+        # TODO: Validate stops with BaseModel
+
+        _, geom_start, _ = geocoding_airport(start)
+        _, geom_dest, _ = geocoding_airport(destination)
+
+        distance = haversine(geom_start[1], geom_start[0], geom_dest[1], geom_dest[0])
+        return _apply_detour(distance, transportation_mode)
+
+    if transportation_mode == "ferry":
+        # Stops are formatted like {"locality":<city>, "county":<country>}
+        # TODO: Validate stops with BaseModel
+
+        _, _, geom_start, _ = geocoding_structured(start)
+        _, _, geom_dest, _ = geocoding_structured(destination)
+        # compute great circle distance between airports
+        distance = haversine(geom_start[1], geom_start[0], geom_dest[1], geom_dest[0])
+
+        return _apply_detour(distance, transportation_mode)
